Putting BVH construction into BuilderStrategy
Renamed Variables

Changed naming convention to snake case.
Renamed 'BVHFlatNode' to just 'Node'
Renamed 'flatTree' to 'nodes'

Created 'BuildStack' class

This class encapsulates the behavior seen in the BVH construction.
It takes some of the complexity out of the build loop, without taking
anything away from the performance.

Laid out new framework for BVH construction

No longer storing 'leaf_size' in BVH structure

Added Node::isLeaf()

Making leaf_size constant, for now

This is simply to reduce the number of parameters
in the build call. Since the build functionality
is getting moved to another class, this can be added
again at a later point.

Added Iterable class

Useful as a drop-in replacement to std::vector.
This is the container used to pass primitives to the BVH builder.

Added 'Config.h'

Used to detect compiler settings.
Possibly, in the future, to detect system settings
such as the availability of AVX2, NEON, or SSE.
May also be used to detect headers with _has_include.

Added 'FastBVH.h'

This is for user convenience so they don't have
to find out which header to include. It's not required
to include, and users can still include what they need
to keep build times low.

Moved 'BVH::build()' into other classes

Created the 'BuildStrategy' class, which
represents the algorithm used to build the BVH.
There are two build strategies. The first is just a 'fake'
build strategy that simply builds a single bounding box around all the
primitives. The second is the original BVH building algorithm.
There's a type definition for the default build strategy that 'points'
to the original build algorithm.

Added BuildStrategy-0 tests.

There may be more tests added here as new algorithms get developed.

Updated the rest of the code base to use BuildStrategy

Updated Makefile

Updated documentation